finalization part 1
