+++ commands/highscores.py
@@ -36,24 +36,6 @@ class Highscores(commands.Cog):
 
             self.client.add_command(outer_cmd(highscore))
 
-    def makeTop10Commands(self):
-        somelist = [RichestClans, BestClans]
-        for highscore in somelist:
-            highscore = highscore()
-
-            def outer_cmd(score: Highscore) -> Command:
-                @commands.command(name=score.NAME)
-                async def cmd(ctx, clanname=None):
-                    if clanname is None and ((clanname := await self.getdefaultclanname(ctx, comment=False)) is None):
-                        clanname = ""
-                    values = score.getDbValues(query="SELECT * FROM {0} WHERE rank < 10 or name = ?".format(score.NAME),
-                                               clan=clanname.lower())
-                    resultmessages = tablify(score.LAYOUT, values)
-                    for i in resultmessages:
-                        await ctx.send(i)
-                return cmd
-            self.client.add_command(outer_cmd(highscore))
-
     @commands.command(name="getplayer")
     async def getplayer(self, ctx: Context, username: str):
         """
